fix: add plugins from the main context to have consistent loading order
Resolves #874
